Move PinListFileSystemMasterClient into GetPinnedFileIdsBench as inner class
